+++ Sources/UI/Login/LoginProcessor.swift
@@ -1,0 +1,17 @@
+//
+//  LoginProcessor.swift
+//  Snabble
+//
+//  Created by Uwe Tilemann on 16.10.22.
+//
+
+import Foundation
+import Combine
+
+protocol LoginProcessing {
+    var loginModel: Loginable? { get }
+
+    func login()
+    func save() async throws
+    func remove()
+}